Avoid kapp phantom diff on AKS
AKS injects an additional facet into the namespaceSelector for mutating
and validating webhook configurations. The next time kapp deploys the
dist, it removes the injected values as drift, and so on in a loop.

We can disable the AKS behavior by specifying an annotation
`admissions.enforcer/disabled: "true"`. This makes the webhooks behave
like on other k8s distributions.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>