Replace webhook cert for unit tests
The pervious cert used sha1 which is no longer accepted by go 1.18.
Regenerating the certs using the same flow we use at runtime in CI. The
new certs are valid for 10 years. These certificates should only be used
for the unit tests and not installed on a real server since the private
key for the cert is public.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>