Publish releases to ghcr (#51)
With the release bundles copied to a public registry, we can also export
vanilla yaml rendered from the bundle with these images. Users can
consume this yaml directly to install the components instead of dealing
with the bundle directly. A Carvel Package resource is also defined
referencing the public bundle.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>