Fail build if grype fails
Grype broke as a result of changes to the ko config as we were asking it
to scan an image that didn't exist. Now we scan all images included in
the bundle, not just images we build. Fail the build if Grype is
unable to complete a scan.

The presence of CVEs will not fail the build.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>